tgbot/requirements.txt
google-cloud-functions
google-cloud-firestore
requests
